Don't attempt to guess whether to run the step if the data the conditions use are not as expected
Add runners which allow the user to choose what happens to the build and whether to run the build step in the event of a failure to evaluate a run condition.
Clean the conditions by removing any code that checks for bad data and makes this decision itself.
Ditch checked exceptions for RunConditions.