Add alternative constructor for StatusCondition
Added alternative constructor for Status condition for use from other plugins
in cases they need to uses it as default. Removes the possibility of messing
up the string in these cases.
Added test cases for the main build states.
Defined a fixed version of the EasyMock to be used in testing.