Test with Java 21 (#94)
Java 21 released Sep 19, 2023. We'd like to announce full support for
Java 21 in early October and would like the most used plugins to be
compiling and testing with Java 21.

The acceptance test harness and plugin bill of materials tests are
already passing with Java 21. This is a further step to improve plugin
readiness for use with Java 21 and for development with Java 21.

The change intentionally tests only two Java configurations, Java 17
and Java 21 because I believe that the risk of a regression that only
affects Java 11 is very low. We generate Java 11 byte code with the
Java 17 and the Java 21 builds, so we're already testing Java 11 byte
code.

The change also uses the same Jenkinsfile as the plugin archetype so
that we won't need to make extra changes when the Jenkins minimum
version is increased beyond 2.409.